auth_service/requirements.txt


Flask==2.2.3
Werkzeug==2.2.3
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0